+++ src/Data/KeyMapping.hpp
@@ -16,6 +16,7 @@ namespace Data {
     public:
         KeyMapping();
         explicit KeyMapping(std::unordered_map<Button, sf::Keyboard::Key> button_to_key);
+        explicit KeyMapping(std::unordered_map<sf::Keyboard::Key, Button> button_to_key);
         void set_button_to_key(const Button& button, const sf::Keyboard::Key& key);
         std::optional<Button> key_to_button(const sf::Keyboard::Key& key);
         std::optional<sf::Keyboard::Key> button_to_key(const Button& button);