+++ src/Toolkit/NormalizedOrigin.hpp
@@ -8,6 +8,12 @@ namespace Toolkit {
         auto bounds = s.getGlobalBounds();
         s.setOrigin(bounds.left+x*bounds.width, bounds.top+y*bounds.height);
     }
+
+    template<class T>
+    void set_local_origin_normalized(T& s, float x, float y) {
+        auto bounds = s.getLocalBounds();
+        s.setOrigin(bounds.left+x*bounds.width, bounds.top+y*bounds.height);
+    }
     
     template<class T>
     void set_origin_normalized_no_position(T& s, float x, float y) {